requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
diff-match-patch==20230430
dj-database-url==0.5.0
Django==3.2.20
django-allauth==0.41.0
django-countries==7.5.1
django-crispy-forms==2.0
django-import-export==3.2.0
et-xmlfile==1.1.0
gunicorn==21.2.0
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
psycopg2==2.9.7
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
stripe==6.2.0
tablib==3.5.0
xlrd==2.0.1
xlwt==1.3.0
